encrypt: content HMAC as IV for stable ciphertext
Random IVs confuse Unison, because multiple reads of the same file show
different content to Unison. Instead, use an HMAC of the file content
as IV, providing stable ciphertext. The GCM requirement of not reusing
IVs is still given, as the only way (apart from HMAC collisions) an IV
is reused is if the file content is the same.